Search fixed and moved to sprints
