+++ src/components/FilterSprint.svelte
@@ -1,17 +1,23 @@
 <script>
     import TaskList from "../components/TaskList.svelte"
 
+
     import { removeDuplicates } from "../../public/js/getSemesterTitles.js"
 
 	export let semester
 	export let displayTaskList
 
+
 	let sprintTitles = []
     let sprintTasks = []
+
 	let showSprints = false
+
+
 
 	// Create a list of sprints, with duplicates removed, to be rendered
     function loadSprints(semester, taskList) {
+		console.log(taskList)
         sprintTasks = []
         sprintTitles = []
         sprintTasks = taskList.filter(task => task.semester === semester)
@@ -22,14 +28,16 @@
 	}
 
 	loadSprints(semester, displayTaskList)
+	
 </script>
 {#if showSprints}
 	{#each sprintTitles as sprint}
 		<details>
-			<summary>{ sprint }</summary>
-			<div>
-				<TaskList bind:sprint bind:sprintTasks />
-			</div>
+			<summary>{ sprint } </summary>
+			
+			
+				<TaskList bind:sprint bind:sprintTasks bind:displayTaskList/>
+			
 		</details>
 	{/each}
 {/if}
@@ -38,21 +46,5 @@
 	details {
 		margin: 1.25em 0;
 	}
-    div {
-		padding: 0 1em;
-		margin-bottom: 1em;
-	}
-	@media (min-width: 40em) {
-		div{
-			display: grid;
-			grid-template-columns: 1fr 1fr;
-			grid-gap: 1em;
-		}
-	}
-
-	@media (min-width: 60em) {
-		div {
-			grid-template-columns: 1fr 1fr 1fr;
-		}
-	}
+    
 </style>
