+++ src/components/TaskSearch.svelte
@@ -1,8 +1,11 @@
 <script>
     import {createEventDispatcher} from "svelte"
+
     const dispatcher = createEventDispatcher()
+
     export let searchTerm
     export let taskTitles
+    
 </script>
 
 <form action="#">
@@ -33,8 +36,9 @@
 <style>
     form {
         display:flex;
-        justify-content: flex-end;
+        justify-content: center;
         align-items: center;
+        margin-bottom: 1em;
     }
     label {
         color: white;
